MRCP session can now optionally have persistent lifetime across a single call for all the applications provided by the module app_unimrcp.
Reworked the use of datastores, which are currently referenced in a similar manner by all the applications such as MRCPRecog(), MRCPSynth() and SynthandRecog().
Added a new option 'plt' to all the applications, which allows to have persistent lifetime, where an MRCP session is created on demand, reused and destroyed on hang-up.
Added a new option 'dse' to all the applications, which allows to specify a particular datastore entry. There can be multiple datastore entries used in the scope of the same call. If the parameter is not set, then the default datastore entry is used.

All the changes are made to be backward compatible. Nothing has to be changed in the existing applications. Everyone is encouraged to verify the behavior, though.